backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-ollama==0.2.3
langchain-chroma==0.2.2
langchain-community==0.3.14
chromadb==0.6.3
pydantic==2.10.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.20
sse-starlette==2.2.1
httpx==0.28.1
greenlet>=3.0.0
unstructured==0.16.12
pypdf==5.1.0
cryptography>=44.0.0
